+++ runtime/service/src/test/java/org/apache/polaris/service/catalog/iceberg/AbstractLocalIcebergCatalogViewTest.java
@@ -222,7 +222,8 @@ public abstract class AbstractLocalIcebergCatalogViewTest
             storageAccessConfigProvider,
             fileIOFactory,
             polarisEventDispatcher,
-            eventMetadataFactory);
+            eventMetadataFactory,
+            new TableMetadataCache(TestTableMetadataCacheConfiguration.withMaxBytes(1024 * 1024)));
     Map<String, String> properties =
         ImmutableMap.<String, String>builder()
             .put(CatalogProperties.FILE_IO_IMPL, "org.apache.iceberg.inmemory.InMemoryFileIO")
@@ -336,7 +337,8 @@ public abstract class AbstractLocalIcebergCatalogViewTest
             storageAccessConfigProvider,
             spiedFactory,
             polarisEventDispatcher,
-            eventMetadataFactory);
+            eventMetadataFactory,
+            new TableMetadataCache(TestTableMetadataCacheConfiguration.withMaxBytes(1024 * 1024)));
     spiedCatalog.initialize(
         CATALOG_NAME,
         ImmutableMap.<String, String>builder()